+++ frontend/next.config.js
@@ -18,7 +18,4 @@ module.exports = withPWA({
   output: "standalone", env: {
     VERSION: version,
   },
-  serverRuntimeConfig: {
-    apiURL: process.env.API_URL ?? 'http://localhost:8080',
-  },
 });